Gravity + Jump

+++ entity/Entity.java
@@ -10,7 +10,7 @@ public abstract class Entity {
     protected int width, height;
     protected Rectangle2D.Float hitbox;
 
-	public Entity(int x, int y, int width, int height) {
+	public Entity(float x, float y, int width, int height) {
 		this.x = x;
 		this.y = y;
         this.width = width;
@@ -22,7 +22,7 @@ public abstract class Entity {
         g.drawRect((int) hitbox.x, (int) hitbox.y, (int) hitbox.width, (int) hitbox.height);
     }
 
-    protected void initHitbox(int x, int y, int width, int height){
+    protected void initHitbox(float x, float y, int width, int height){
         hitbox = new Rectangle2D.Float(x, y, width, height);
     }
